# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para obter temperatura com base no CEP usando Swagger.",
        "title": "Zipcode Temperature API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/temperature/{cep}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "CEP para buscar a temperatura",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TemperatureResponse"
                        }
                    },
                    "404": {
                        "description": "can not find zipcode",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "invalid zipcode",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.TemperatureResponse": {
            "type": "object",
            "properties": {
                "temp_C": {
                    "type": "number"
                },
                "temp_F": {
                    "type": "number"
                },
                "temp_K": {
                    "type": "number"
                }
            }
        }
    }
}